move test collector and test suite
