+++ protostar/commands/test/test_command.py
@@ -4,7 +4,6 @@ from typing import List, Optional
 
 from protostar.cli.activity_indicator import ActivityIndicator
 from protostar.cli.command import Command
-from protostar.commands.test.test_collector import TestCollector
 from protostar.commands.test.test_collector_summary_formatter import (
     format_test_collector_summary,
 )
@@ -16,6 +15,7 @@ from protostar.commands.test.testing_live_logger import TestingLiveLogger
 from protostar.commands.test.testing_seed import determine_testing_seed
 from protostar.commands.test.testing_summary import TestingSummary
 from protostar.compiler import ProjectCairoPathBuilder
+from protostar.test_runner import TestCollector
 from protostar.utils.compiler.pass_managers import (
     StarknetPassManagerFactory,
     TestCollectorPassManagerFactory,
