+++ protostar/commands/test/test_config.py
@@ -6,7 +6,7 @@ from starkware.starknet.testing.contract import StarknetContract
 from typing_extensions import Self
 
 from protostar.commands.test.fuzzing.strategy_descriptor import StrategyDescriptor
-from protostar.commands.test.test_suite import TestCase
+from protostar.test_runner.test_suite import TestCase
 from protostar.commands.test.testing_seed import random_seed, Seed
 from protostar.protostar_exception import ProtostarException
 from protostar.utils.abi import has_function_parameters
