+++ protostar/commands/test/test_shared_tests_state.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 from protostar.commands.test.test_results import PassedTestCaseResult, TestResult
 
 if TYPE_CHECKING:
-    from protostar.commands.test.test_collector import TestCollector
+    from protostar.test_runner import TestCollector
 
 
 class SharedTestsState:
