+++ protostar/test_runner/__init__.py
@@ -1,0 +1,1 @@
+from .test_collector import TestCollector
