+++ protostar/test_runner/test_collector_test.py
@@ -8,9 +8,10 @@ from starkware.cairo.lang.compiler.preprocessor.preprocessor_error import (
     PreprocessorError,
 )
 
-from protostar.commands.test.test_collector import TestCollector
-from protostar.commands.test.test_suite import TestSuite, TestCase
 from protostar.utils.starknet_compilation import StarknetCompiler
+
+from .test_collector import TestCollector
+from .test_suite import TestCase, TestSuite
 
 
 @pytest.fixture(name="project_root")
