+++ tests/benchmarks/test_benchmark.py
@@ -1,7 +1,7 @@
-import multiprocessing
-from contextlib import asynccontextmanager
 import asyncio
+import multiprocessing
 import threading
+from contextlib import asynccontextmanager
 from multiprocessing.managers import SyncManager
 from pathlib import Path
 from string import Template
@@ -11,13 +11,13 @@ import pytest
 from starkware.starknet.services.api.contract_class import ContractClass
 
 from protostar.commands.test.starkware.test_execution_state import TestExecutionState
-from protostar.commands.test.test_collector import TestCollector
 from protostar.commands.test.test_config import TestConfig
 from protostar.commands.test.test_runner import TestRunner
 from protostar.commands.test.test_shared_tests_state import SharedTestsState
-from protostar.commands.test.test_suite import TestSuite, TestCase
+from protostar.test_runner import TestCollector
+from protostar.test_runner.test_suite import TestCase, TestSuite
 from protostar.utils.compiler.pass_managers import ProtostarPassMangerFactory
-from protostar.utils.starknet_compilation import StarknetCompiler, CompilerConfig
+from protostar.utils.starknet_compilation import CompilerConfig, StarknetCompiler
 from tests.benchmarks.constants import ROUNDS_NUMBER
 
 SCRIPT_DIRECTORY = Path(__file__).parent
